Rename projects to content sources
