--- lib/glossia/builds.ex
@@ -1,16 +1,0 @@
-defmodule Glossia.Builds do
-  @moduledoc false
-
-  alias Glossia.Builds.BuildWorker
-  require Logger
-
-  def trigger_build(attrs) do
-    attrs
-    |> BuildWorker.new()
-    |> Oban.insert()
-    |> case do
-      {:ok, _} -> :ok
-      {:error, error} -> {:error, error}
-    end
-  end
-end
