+++ lib/glossia/localizations.ex
@@ -1,9 +1,7 @@
 defmodule Glossia.Localizations do
   @moduledoc false
 
-  alias Glossia.Localizations.Workers.LocalizeWorker
-  alias Glossia.Projects
-  alias Glossia.Projects.Project
+  # alias Glossia.Localizations.Workers.LocalizeWorker
   @behaviour Glossia.Authorization.Policy
 
   # Types
@@ -21,33 +19,34 @@ defmodule Glossia.Localizations do
           localization :: any(),
           opts :: process_localization_opts()
         ) :: :ok | {:error, term()}
-  def process_localization(localization, %{project_id: project_id} = _opts) do
-    project = Projects.find_project_by_id(project_id)
+  def process_localization(_localization, %{project_id: _project_id} = _opts) do
+    :ok
+    # project = Projects.find_project_by_id(project_id)
 
-    content_platform_module =
-      Glossia.ContentSources.get_platform_module(project.content_platform)
+    # content_platform_module =
+    #   Glossia.ContentSources.get_platform_module(project.content_platform)
 
-    version = localization.version
+    # version = localization.version
 
-    unique_id =
-      case content_platform_module.get_content_branch_id(
-             project.id_in_content_platform,
-             %{version: version}
-           ) do
-        nil -> version
-        id -> id
-      end
+    # unique_id =
+    #   case content_platform_module.get_content_branch_id(
+    #          project.id_in_content_platform,
+    #          %{version: version}
+    #        ) do
+    #     nil -> version
+    #     id -> id
+    #   end
 
-    %{localization: localization, project_id: project_id, unique_id: unique_id}
-    |> LocalizeWorker.new(replace: [:args])
-    |> Oban.insert()
-    |> case do
-      {:ok, _} -> :ok
-      {:error, error} -> {:error, error}
-    end
+    # %{localization: localization, project_id: project_id, unique_id: unique_id}
+    # |> LocalizeWorker.new(replace: [:args])
+    # |> Oban.insert()
+    # |> case do
+    #   {:ok, _} -> :ok
+    #   {:error, error} -> {:error, error}
+    # end
   end
 
-  def authorize(:create, %Project{} = authenticated_project, %Project{} = project) do
-    if authenticated_project.id == project.id, do: :ok, else: :error
+  def authorize(_action, _subject, _object) do
+    :ok
   end
 end
