+++ lib/glossia_web/controllers/api/localization_controller.ex
@@ -1,7 +1,7 @@
 defmodule GlossiaWeb.Controllers.API.LocalizationController do
   @moduledoc false
 
-  alias Glossia.Localizations
+  # alias Glossia.Localizations
   use GlossiaWeb.Helpers.App, :controller
   use OpenApiSpex.ControllerSpecs
 
@@ -9,7 +9,7 @@ defmodule GlossiaWeb.Controllers.API.LocalizationController do
 
   tags ["localizations"]
 
-  alias Glossia.Localizations, as: Localizations
+  # alias Glossia.Localizations, as: Localizations
   alias GlossiaWeb.APISchemas.Localizations.CreateResponse
   alias GlossiaWeb.APISchemas.Localizations.Localization
 
@@ -23,24 +23,24 @@ defmodule GlossiaWeb.Controllers.API.LocalizationController do
 
   @spec create(conn :: Plug.Conn.t(), any) :: Plug.Conn.t()
   @dialyzer {:nowarn_function, create: 2}
-  def create(conn, _params) do
-    Glossia.Authorization.permit!(
-      Glossia.Localizations,
-      :create,
-      GlossiaWeb.Auth.authenticated_subject(conn),
-      GlossiaWeb.LiveViewMountablePlug.url_project(conn)
-    )
+  def create(_conn, _params) do
+    # Glossia.Authorization.permit!(
+    #   Glossia.Localizations,
+    #   :create,
+    #   GlossiaWeb.Auth.authenticated_subject(conn),
+    #   GlossiaWeb.LiveViewMountablePlug.url_project(conn)
+    # )
 
-    %{body_params: %Localization{} = localization} = conn
+    # %{body_params: %Localization{} = localization} = conn
 
-    result =
-      Localizations.process_localization(localization, %{
-        project_id: conn.assigns[:authenticated_project].id
-      })
+    # result =
+    #   Localizations.process_localization(localization, %{
+    #     project_id: conn.assigns[:authenticated_project].id
+    #   })
 
-    case result do
-      :ok -> conn |> put_status(:ok) |> json(%CreateResponse{})
-      {:error, _error} -> conn |> put_status(:internal_server_error) |> json(%CreateResponse{})
-    end
+    # case result do
+    #   :ok -> conn |> put_status(:ok) |> json(%CreateResponse{})
+    #   {:error, _error} -> conn |> put_status(:internal_server_error) |> json(%CreateResponse{})
+    # end
   end
 end
