+++ lib/glossia_web/live/projects/versions.ex
@@ -6,10 +6,11 @@ defmodule GlossiaWeb.LiveViews.Projects.Versions do
     {:ok, socket}
   end
 
-  def get_versions(socket) do
-    project = GlossiaWeb.LiveViewMountablePlug.url_project(socket)
-    content_platform_module = Glossia.ContentSources.get_platform_module(project.content_platform)
-    versions = content_platform_module.get_versions(project.id_in_content_platform)
-    versions
+  def get_versions(_socket) do
+    # project = GlossiaWeb.LiveViewMountablePlug.url_project(socket)
+    # content_platform_module = Glossia.ContentSources.get_platform_module(project.content_platform)
+    # versions = content_platform_module.get_versions(project.id_in_content_platform)
+    # versions
+    []
   end
 end
