--- test/glossia/projects/models/project_token_test.exs
@@ -1,33 +1,0 @@
-defmodule Glossia.Projects.ProjectTokenTest do
-  use Glossia.DataCase
-
-  import Glossia.ProjectsFixtures
-  import Glossia.Projects.ProjectToken
-
-  describe "generate_token" do
-    test "generates the token successfully" do
-      # Given
-      project = project_fixture()
-
-      # When
-      {:ok, token, _claims} = project.id |> generate_token_for_project_with_id()
-
-      # Then
-      assert token != ""
-    end
-  end
-
-  describe "get_project_id_from_token" do
-    test "it returns the build_id from a generated token" do
-      # Given
-      project = project_fixture()
-
-      # When
-      {:ok, token, _claims} = project.id |> generate_token_for_project_with_id()
-      {:ok, project_id} = token |> get_project_id_from_token()
-
-      # Then
-      assert project_id == project.id
-    end
-  end
-end
